Moved more things around for release
