+++ go/libraries/doltcore/sqle/logictest/testdata/resultlog.txt
